Begin separation of Arim API from ArimBansLib
