+++ ArimBans3/src/main/java/space/arim/bans/internal/backend/subjects/Subjects.java
@@ -26,7 +26,8 @@ import space.arim.bans.api.Punishment;
 import space.arim.bans.api.PunishmentType;
 import space.arim.bans.api.Subject;
 import space.arim.bans.api.exception.InvalidUUIDException;
-import space.arim.bans.api.util.minecraft.MinecraftUtil;
+
+import space.arim.api.util.minecraft.MinecraftUtil;
 
 public class Subjects implements SubjectsMaster {
