+++ ArimBans3/src/main/java/space/arim/bans/internal/frontend/format/Formats.java
@@ -37,7 +37,8 @@ import space.arim.bans.api.CommandType.SubCategory;
 import space.arim.bans.api.exception.InternalStateException;
 import space.arim.bans.api.exception.InvalidSubjectException;
 import space.arim.bans.api.exception.PlayerNotFoundException;
-import space.arim.bans.api.util.StringsUtil;
+
+import space.arim.api.util.StringsUtil;
 
 public class Formats implements FormatsMaster {
