+++ ArimBans3/src/main/java/space/arim/bans/internal/sql/Sql.java
@@ -32,11 +32,12 @@ import com.zaxxer.hikari.HikariDataSource;
 
 import space.arim.bans.ArimBans;
 import space.arim.bans.api.exception.InternalStateException;
-import space.arim.bans.api.sql.ExecutableQuery;
-import space.arim.bans.api.util.StringsUtil;
 import space.arim.bans.internal.sql.SqlQuery.Query;
 
 import space.arim.universal.util.collections.CollectionsUtil;
+
+import space.arim.api.sql.ExecutableQuery;
+import space.arim.api.util.StringsUtil;
 
 public class Sql implements SqlMaster {
