+++ ArimBans3/src/main/java/space/arim/bans/internal/sql/SqlMaster.java
@@ -21,9 +21,10 @@ package space.arim.bans.internal.sql;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-import space.arim.bans.api.sql.ExecutableQuery;
 import space.arim.bans.internal.Component;
 import space.arim.bans.internal.sql.SqlQuery.Query;
+
+import space.arim.api.sql.ExecutableQuery;
 
 public interface SqlMaster extends Component {
 	@Override
