+++ ArimBansExtended/src/main/java/space/arim/bans/extended/ArimBansExtendedBungee.java
@@ -23,9 +23,10 @@ import java.util.Set;
 
 import net.md_5.bungee.api.plugin.Plugin;
 
-import space.arim.bans.api.util.minecraft.bungee.BungeeUtil;
 import space.arim.bans.extended.bungee.CommandSkeleton;
 import space.arim.bans.extended.bungee.SignInterceptorProtocolize;
+
+import space.arim.api.util.minecraft.bungee.BungeeUtil;
 
 public class ArimBansExtendedBungee extends Plugin implements ArimBansExtendedPluginBase {
