+++ ArimBansLib/src/main/java/space/arim/api/sql/SQLPlugin.java
@@ -16,30 +16,19 @@
  * along with ArimBansLib. If not, see <https://www.gnu.org/licenses/>
  * and navigate to version 3 of the GNU General Public License.
  */
-package space.arim.bans.api.util.web;
+package space.arim.api.sql;
 
-public class GeoIpInfo {
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public interface SQLPlugin {
 	
-	public final String address;
-	public final String country_code;
-	public final String country_name;
-	public final String region_code;
-	public final String region_name;
-	public final String city;
-	public final String zip;
-	public final double latitude;
-	public final double longitude;
+	void executionQueries(ExecutableQuery...queries) throws SQLException;
 	
-	GeoIpInfo(String address, String country_code, String country_name, String region_code, String region_name, String city, String zip, double latitude, double longitude) {
-		this.address = address;
-		this.country_code = country_code;
-		this.country_name = country_name;
-		this.region_code = region_code;
-		this.region_name = region_name;
-		this.city = city;
-		this.zip = zip;
-		this.latitude = latitude;
-		this.longitude = longitude;
-	}
+	void executionQuery(String query, Object...parameters) throws SQLException;
+	
+	ResultSet[] selectionQueries(ExecutableQuery...queries) throws SQLException;
+	
+	ResultSet selectionQuery(String query, Object...parameters) throws SQLException;
 	
 }
