+++ ArimBansLib/src/main/java/space/arim/api/util/StringsUtil.java
@@ -16,7 +16,7 @@
  * along with ArimBansLib. If not, see <https://www.gnu.org/licenses/>
  * and navigate to version 3 of the GNU General Public License.
  */
-package space.arim.bans.api.util;
+package space.arim.api.util;
 
 import java.text.SimpleDateFormat;
 import java.util.Arrays;
