+++ ArimBansLib/src/main/java/space/arim/api/util/minecraft/MinecraftUtil.java
@@ -16,7 +16,7 @@
  * along with ArimBansLib. If not, see <https://www.gnu.org/licenses/>
  * and navigate to version 3 of the GNU General Public License.
  */
-package space.arim.bans.api.util.minecraft;
+package space.arim.api.util.minecraft;
 
 import java.util.ArrayList;
